backend/requirements.txt
fastapi
uvicorn
tweepy
python-dotenv